+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 ## 1.0.1
  - Privacy related issue fixed.
 ## 1.0.2
-- Json formatted in response body..
+- Json formatted in response body.
+## 1.0.3
+- Local push notification added for request,response and error.